requirements-ci.txt
aioamqp==0.10.0
asynctest==0.9.0
flake8==3.3.0
flake8-import-order==0.12.0
msgpack-python==0.4.8
pytest==3.1.0
pytest-asyncio==0.6.0
pytest-mock==1.6.0
SQLAlchemy==1.1.9
